requirements.txt
torchvision>=0.13.1+cu116
torch>=1.12.1+cu116
Pillow>=7.0.0
numpy>=1.21.5
runx>=0.0.11













